Add a Vagrant config
The config is actually auto generated by:
  vagrant init archlinux/archlinux
Writing commands into a vanilla ISO running in QEMU is annoying since it
would have to be done through the monitor which supports sending only
one key at a time. A Vagrant box is not quite the same thing as running
an ISO but it skips most of the manual steps.