Vagrantfile: Increase RAM to 4GiB
The archlinux vagrant vox uses a swapfile by default and the limited RAM
in the box causes aconfmgr to fail when it tries to guess system
state by reflecting it in a /tmp/ directory.

To avoid issues with /tmp/ being out of memory in cases other then
swapfile increase RAM available to this vagrant box.